Fix/Create self links, if not present or invalid, support column white-list for conversion

+++ src/stactools/geoparquet_items/commands.py
@@ -6,7 +6,7 @@ import os
 # Used only for partitioning / dask variant
 import pathlib
 import shutil
-from typing import Any, Optional, Sequence
+from typing import Any, Dict, Optional, Sequence
 
 import click
 import dask.bag as db
@@ -18,11 +18,19 @@ import pyogrio
 import requests
 import stac_geoparquet
 from click import Command, Group
+from urllib.parse import urlparse, urljoin
 
 logger = logging.getLogger(__name__)
 
-IGNORE_FIELDS = ["stac_version", "type", "assets"]
+STAC_ITEM_TYPES = ["application/json", "application/geo+json"]
+SELF_LINK_COLUMN = "self_link"
 
+def is_self_link(link: Dict[str, Any]) -> bool:
+    return (
+        link["rel"] == "self"
+        and (not link["type"] or link["type"] in STAC_ITEM_TYPES)
+        and urlparse(link["href"]).netloc
+    )
 
 def create_geoparquetitems_command(cli: Group) -> Command:
     """Creates the stactools-geoparquet-items command line utility."""
@@ -55,12 +63,19 @@ def create_geoparquetitems_command(cli: Group) -> Command:
         help="Tries to add the absolute link to the source STAC Item to a column named 'self_link'",
         is_flag=True,
     )
+    @click.option(
+        "--baseurl",
+        default=None,
+        help="If provided, tries to fix invalid self URLs by mapping to the source folder to the given URL. "
+        + "Only applies if the given source is a local folder. Must have a slash at the end."
+    )
     def create_command(
         source: str,
         destination: str,
         collection: str = "",
         partition: int = 1,
         selflink: bool = False,
+        baseurl: str = None
     ) -> None:
         """Create geoparquet from STAC Items
 
@@ -68,6 +83,9 @@ def create_geoparquetitems_command(cli: Group) -> Command:
             source (str): Link to a list of STAC Items (ItemCollection) or a folder with STAC files.
             destination (str): Path where the geoparquet file will be stored.
         """
+        if baseurl is not None and not baseurl.endswith("/"):
+            raise Exception("baseurl must end with a slash")
+
         p = pathlib.Path(destination)
         if p.is_dir():
             shutil.rmtree(p, ignore_errors=True)
@@ -105,14 +123,43 @@ def create_geoparquetitems_command(cli: Group) -> Command:
 
             print("Found {} potential STAC Items".format(len(paths)))
 
+            def load_file(p: pathlib.Path):
+                with p.open() as f:
+                    stac = json.load(f)
+                    if baseurl:
+                        # remove self link
+                        for link in stac["links"]:
+                            if is_self_link(link):
+                                stac["links"].remove(link)
+                                break
+
+                        # create self link
+                        rel_path = str(p.relative_to(source)).replace('\\', '/')
+                        href = urljoin(baseurl, rel_path)
+                        self_link = {
+                            'href': href,
+                            'type': STAC_ITEM_TYPES[0],
+                            'rel': 'self'
+                        }
+
+                        stac["links"].append(self_link)
+
+                    return stac
+
             if partition > 1:
                 bag = (
                     db.from_sequence(paths, npartitions=partition)
-                    .map(lambda file: file.read_text())
-                    .map(json.loads)
-                    .filter(lambda item: item["type"] == "Feature")
+                    .map(load_file)
+                    .filter(lambda stac: stac["type"] == "Feature")
                 )
-
+            else:
+                for p in paths:
+                    stac = load_file(p)
+                    if stac["type"] == "Feature":
+                        items.append(stac)
+            
+            del paths
+        
         def create_fn(items: Sequence[dict[str, Any]]) -> geopandas.GeoDataFrame:
             return stac_geoparquet.to_geodataframe(items, add_self_link=selflink)
 
@@ -136,13 +183,6 @@ def create_geoparquetitems_command(cli: Group) -> Command:
             df.to_parquet(destination, schema=schema, write_index=False)
             print("Wrote geoparquet file(s)")
         else:
-            for p in paths:
-                with p.open() as f:
-                    item = json.load(f)
-                    if item["type"] == "Feature":
-                        items.append(item)
-            del paths
-
             num = len(items)
             if num > 0:
                 print(f"Loaded {num} actual STAC Items")
@@ -184,10 +224,14 @@ def create_geoparquetitems_command(cli: Group) -> Command:
     @click.option(
         "--exclude",
         "-e",
-        default=",".join(IGNORE_FIELDS),
-        show_default=True,
-        help="A list of comma-separated fields that should be excluded from the target file. "
-        + " Use 'none' to include all fields."
+        default=None,
+        help="A list of comma-separated fields that should be excluded from the target file."
+    )
+    @click.option(
+        "--include",
+        "-i",
+        default=None,
+        help="A list of comma-separated fields that should be included in the target file. This option takes preference over `exclude`."
     )
     @click.option(
         "--format",
@@ -205,6 +249,7 @@ def create_geoparquetitems_command(cli: Group) -> Command:
         destination: str,
         format: str = "gpkg",
         exclude: Optional[str] = None,
+        include: Optional[str] = None,
     ) -> None:
         """Convert geoparquet to other OGR file formats
 
@@ -215,19 +260,16 @@ def create_geoparquetitems_command(cli: Group) -> Command:
         if not os.path.exists(source):
             raise Exception("Source file does not exist")
 
-        if exclude is None:
-            to_exclude = IGNORE_FIELDS
-        elif exclude == "none" or exclude == "NONE":
-            to_exclude = []
-        else:
-            to_exclude = exclude.split(",")
-
         columns = None
-        if len(to_exclude) > 0:
-            schema = pq.read_schema(source)
-            columns = schema.names.copy()
-            for col in to_exclude:
-                columns.remove(col.strip())
+        if include is not None:
+            columns = [col.strip() for col in include.split(",")]
+        elif exclude is not None:
+            to_exclude = exclude.split(",")
+            if len(to_exclude) > 0:
+                schema = pq.read_schema(source)
+                columns = schema.names.copy()
+                for col in to_exclude:
+                    columns.remove(col.strip())
 
         df = geopandas.read_parquet(source, columns=columns)
         pyogrio.write_dataframe(df, destination, driver=format)
